requirements.txt
openai>=1.0.0
streamlit>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pygame>=2.5.0
python-slugify>=8.0.0
attrs>=23.1.0
